Split Arena in PerioDicRunner and RequestHandler
